refactor: rename 'services' to 'business'
